# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carteira Financeira API",
        "description": "API para gerenciamento de carteira financeira",
        "contact": {
            "name": "Suporte API",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/cadastrar": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registra um novo usuário no sistema",
                "description": "Registra um novo usuário.",
                "operationId": "9cfafe40f38f40316cbbbe9452a5517c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone",
                                    "document"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "12345678901"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "123e4567-e89b-12d3-a456-426614174000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@email.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "11999999999"
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "12345678901"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "O email já está em uso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autentica um usuário e retorna o token de acesso",
                "description": "Autentica o usuário e cria token.",
                "operationId": "64c17bd97d94fdaf7e5fb4f4b0a6f688",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "123e4567-e89b-12d3-a456-426614174000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@email.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "11999999999"
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "12345678901"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abcdef123456..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Credenciais inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/perfil": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Retorna os dados do usuário autenticado, incluindo os dados da conta",
                "description": "Obtém o usuário autenticado.",
                "operationId": "78bc427fb1a9ae6e3602dd4d7c082b69",
                "responses": {
                    "200": {
                        "description": "Dados do usuário retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "123e4567-e89b-12d3-a456-426614174000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@email.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "11999999999"
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "12345678901"
                                                },
                                                "conta": {
                                                    "description": "Dados da conta vinculada ao usuário",
                                                    "properties": {
                                                        "id": {
                                                            "description": "ID público da conta",
                                                            "type": "string",
                                                            "example": "c1a2b3c4-d5e6-7890-abcd-1234567890ef"
                                                        },
                                                        "user_id": {
                                                            "description": "ID público do usuário",
                                                            "type": "string",
                                                            "example": "123e4567-e89b-12d3-a456-426614174000"
                                                        },
                                                        "balance": {
                                                            "description": "Saldo atual da conta",
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 1000.5
                                                        },
                                                        "status": {
                                                            "description": "Status atual da conta",
                                                            "type": "string",
                                                            "example": "ativa"
                                                        },
                                                        "created_at": {
                                                            "description": "Data de criação",
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2024-05-01T12:00:00Z"
                                                        },
                                                        "updated_at": {
                                                            "description": "Data de última atualização",
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2024-05-10T15:30:00Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Revoga o token de acesso do usuário atual",
                "description": "Desconecta o usuário (revoga o token).",
                "operationId": "640784fea1cbcc077f936d2f9e623d75",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}